# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agorich Pharma API
  description: |
    Comprehensive API documentation for Agorich Pharma's hyperlocal pharma intelligence operating system.

    ## Overview
    This API provides endpoints for managing the pharmaceutical distribution ecosystem including:
    - AI-powered collection calls via VAPI
    - Inventory management and alerts
    - Invoice processing and approval workflows
    - Real-time metrics and analytics
    - Emergency controls and system management

    ## Authentication
    - **User Authentication**: Supabase Auth with JWT tokens (Cookie-based for web clients)
    - **Agent API Key**: `x-agent-api-key` header for backend-to-backend communication
    - **Admin API Key**: `x-admin-api-key` header for admin operations

    ## Rate Limiting
    - Default: 100 requests per minute per IP
    - AI endpoints: 20 requests per minute
    - Webhook endpoints: 200 requests per minute
  version: 1.0.0
  contact:
    name: Agorich Pharma Tech Team
    email: tech@agorichpharma.com
  license:
    name: Proprietary
servers:
  - url: https://api.agorichpharma.com
    description: Production server
  - url: https://staging-api.agorichpharma.com
    description: Staging server
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/command-center/chat:
    post:
      tags:
        - Command Center
      summary: Chat with AI command center
      description: |
        Natural language interface to the command center. Accepts conversational queries
        and returns AI-powered insights about sales, inventory, and AR data.
      operationId: commandCenterChat
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            example:
              message: "Show me the top 5 retailers with highest overdue payments"
              context:
                region: "Maharashtra"
                timeframe: "30days"
      responses:
        '200':
          description: AI response generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/cron/trigger-vapi-calls:
    post:
      tags:
        - Agent Connect
      summary: Trigger collection calls
      description: |
        Triggers VAPI AI agent calls to all overdue customers.

        This endpoint is typically called by a cron job to initiate collection calls.
      operationId: triggerVapiCalls
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Calls initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerCallsResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-agent-api-key
      description: API key for agent-to-agent communication
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Admin API key for administrative operations
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase Auth
    WebhookSecretAuth:
      type: apiKey
      in: header
      name: x-vapi-webhook-secret
      description: Webhook verification secret
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message description"
      required:
        - success
        - error
...
    ChatRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "Show me the top 5 retailers with highest overdue payments"
        context:
          type: object
          properties:
            region:
              type: string
            timeframe:
              type: string
    ChatResponse:
      type: object
      properties:
        success:
          type: boolean
        response:
          type: string
        data:
          type: object
          description: Structured data if applicable (charts, tables, etc.)
        suggestions:
          type: array
          items:
            type: string
...
    TriggerCallsResponse:
      type: object
      properties:
        success:
          type: boolean
        callsInitiated:
          type: integer
        results:
          type: array
          items:
            type: object
...